Move configuration to a more sensible location
