+++ mailchimp-webhooks/src/main/java/uk/ac/stfc/facilities/mailing/Application.java
@@ -2,7 +2,7 @@ package uk.ac.stfc.facilities.mailing;
 
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.builder.SpringApplicationBuilder;
-import uk.ac.stfc.facilities.mailing.spring.MailchimpApplicationContextInitializer;
+import uk.ac.stfc.facilities.mailing.spring.config.MailchimpApplicationContextInitializer;
 
 /**
  * Entry point for the Mailchimp webhooks application.
