+++ mailchimp-webhooks/src/main/java/uk/ac/stfc/facilities/mailing/spring/config/BisAppSettingsPropertySource.java
@@ -1,4 +1,4 @@
-package uk.ac.stfc.facilities.mailing.spring.environment.config;
+package uk.ac.stfc.facilities.mailing.spring.config;
 
 import org.springframework.core.env.PropertySource;
 import uk.stfc.bisapps.config.BISAppProperties;
